% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_express.R
\name{express_eval.lm}
\alias{express_eval.lm}
\title{Express a linear model evaluation}
\usage{
\method{express_eval}{lm}(
  model,
  orig_data,
  by_factor = NULL,
  by_covar = NULL,
  grid = TRUE,
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  f_pal = "Dynamic",
  cov_pal = "Sunset Dark",
  cov_pal_rev = FALSE,
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{lm}.}

\item{orig_data}{Original data on which the model was built.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size.}

\item{grid}{Whether to return a combined \code{cowplot::plot_grid} object of all
the plots (default = TRUE). If FALSE, will return a list of ggplot objects.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
parameter estimate.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_color_discrete_qualitative} palette.}

\item{cov_pal}{"Covariate" palette. If mapping an aesthetic to a covariate,
which palette to use? Must be a
\code{colorspace::scale_color_continuous_sequential} palette.}

\item{cov_pal_rev}{Reverse the covariate palette? Binary TRUE/FALSE.}

\item{...}{Unused.}
}
\value{
A \code{cowplot::plot_grid} object of the model diagnostics with
mapping to the supplied factor and/or covariate, or if \code{grid = FALSE},
a list of ggplot objects.
}
\description{
Express the evaluations of an \code{lm}  object in a similar way to
\code{plot.lm}, but allow for aestheticizing by factors or
covariates potentially not included in the model definition itself.
}
\details{
Included in the output is:
\enumerate{
\item A linear predictor (fitted) vs. residuals plot. (\link{express_linpred.lm})
\item A Q-Q plot of model residuals. (\link{express_qqresid.lm})
\item A scale-location plot. (\link{express_scloc.lm})
\item A leverage vs. residuals plot. (\link{express_lev.lm})
}
}
