% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_eval}
\alias{express_eval}
\title{Express a model evaluation}
\usage{
express_eval(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A \code{cowplot::plot_grid} object for model diagnostics.
}
\description{
Plot model evaluations as a \code{cowplot::plot_grid} object that includes
aesthetics mapped to multiple variables, whether or not they are included in
the actual definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Plot diagnostics for a (multivariate) linear model. (\link{express_eval.lm})
\item \code{glm}: Plot diagnostics for a (multivariate) generalized linear model.
(\link{express_eval.glm})
\item \code{lmm}: Plot diagnostics for an \code{lme4::lmerMod} model. Supports
linear and random effects. (\link{express_eval.lmerMod})
\item \code{glmm}: Plot diagnostics for an \code{lme4::glmerMod} model. Supports
linear and random effects. (\link{express_eval.glmerMod})
\item \code{gam}: Plot diagnostics for an \code{mgcv::gam} model. (\link{express_eval.gam})
}
}
\details{
Generic plotting of model diagnostics as ggplot objects with optional
additional faceting, within a \code{cowplot::plot_grid}. Uses cowplot and colorspace
along with ggplot2.
}
