% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_express.R
\name{express_fit.gam}
\alias{express_fit.gam}
\title{Express a GAM fit}
\usage{
\method{express_fit}{gam}(
  model,
  orig_data = NULL,
  by_factor = NULL,
  by_covar = NULL,
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  f_pal = "Dynamic",
  cov_pal = "Sunset Dark",
  cov_pal_rev = FALSE,
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{mgcv::gam}.}

\item{orig_data}{Original data on which the model was built. This can either
be supplied with the original data frame directly, or, with \code{mgcv::gam}
objects, the original data can be contained within the model itself by
supplying the argument \code{control = list(keepData = TRUE)}, in which case,
orig_data can be left NULL and the data will be taken from the model object.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
parameter estimate.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_color_discrete_qualitative} palette.}

\item{cov_pal}{"Covariate" palette. If mapping an aesthetic to a covariate,
which palette to use? Must be a
\code{colorspace::scale_color_continuous_sequential} palette.}

\item{cov_pal_rev}{Reverse the covariate palette? Binary TRUE/FALSE.}

\item{...}{Unused.}
}
\value{
A ggplot object of the model fit with
mapping to the supplied factor and/or covariate.
}
\description{
Express the model fit (fitted vs. observed) of an \code{mgcv::gam} object, allowing for aestheticizing by factors or
covariates potentially not included in the model definition itself.
}
