% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_express.R
\name{express_fit.lm}
\alias{express_fit.lm}
\title{Express a linear model fit}
\usage{
\method{express_fit}{lm}(
  model,
  orig_data,
  by_factor = NULL,
  by_covar = NULL,
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  f_pal = "Dynamic",
  cov_pal = "Sunset Dark",
  cov_pal_rev = FALSE,
  ...
)
}
\arguments{
\item{model}{A supplied model of type lm.}

\item{orig_data}{Original data on which the model was built.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
model fit.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_color_discrete_qualitative} palette.}

\item{cov_pal}{"Covariate" palette. If mapping an aesthetic to a covariate,
which palette to use? Must be a
\code{colorspace::scale_color_continuous_sequential} palette.}

\item{cov_pal_rev}{Reverse the covariate palette? Binary TRUE/FALSE.}

\item{...}{Unused.}
}
\value{
A ggplot object of the full model fit
mapped to the supplied factor and/or covariate.
}
\description{
Express the fit of a linear model according to its full residuals, allowing for aestheticizing
by factors or covariates potentially not in the model definition itself.
}
