% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_fit}
\alias{express_fit}
\title{Express a model fit}
\usage{
express_fit(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A ggplot object for the full model fit.
}
\description{
Plot the fit of a full model as a ggplot object that includes
aesthetics mapped to multiple variables, whether or not they are included in
the actual definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Fit for a (multivariate) linear model. (\link{express_fit.lm})
\item \code{glm}: Fit for a (multivariate) generalized linear model. (\link{express_fit.glm})
\item \code{lmm}: Fit for an \code{lme4::lmerMod} model. (\link{express_fit.lmerMod})
\item \code{glmm}: Fit for an \code{lme4::glmerMod} model. (\link{express_fit.glmerMod})
\item \code{gam}: Fit for an \code{mgcv::gam} model. (\link{express_fit.gam})
}
}
\details{
Generic plotting of model fits as ggplot objects with optional
additional faceting. Uses cowplot and colorspace
along with ggplot2.
}
