% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_express.R
\name{express_gauge.lmerMod}
\alias{express_gauge.lmerMod}
\title{Express LMM residuals to gauge systematic effects}
\usage{
\method{express_gauge}{lmerMod}(
  model,
  orig_data,
  by_factor = NULL,
  by_covar = NULL,
  covar_fit = "linear",
  res_type = "pearson",
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  f_pal = "Dynamic",
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{lme4::lmerMod}.}

\item{orig_data}{Original data on which the model was built.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour. Residuals will be tested for systematic variation
using a one-way ANOVA.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size. Residuals will
be tested for systematic variation by regressing the residuals along the
covariate. See \code{covar_fit}.}

\item{covar_fit}{Specify a transform for the covariate for regression.
If there is a presumed relatonship between the covariate
and the residuals, might it be nonlinear? Default is
\code{"linear"} (i.e., no transform) but can be any of \code{c("linear", "exp", "log")}.}

\item{res_type}{Residual type. Will work with any of the types that can be
passed to \code{residuals.default}.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
regression estimate of the covariate and residuals.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_fill_discrete_qualitative} palette.}

\item{...}{Unused.}
}
\value{
A named list containing:
\itemize{
\item \code{"plots"}: If using by_factor, a boxplot of the residuals by factor group.
If using by_covar, a scatterplot of the residuals along the covariate. If
using both, both. Plots are named by the dependent variable of the model and
either the factor or covariate split for easy differentiation.
\item \code{"summaries"} If using by_factor, the summary of the one-way ANOVA for the
residuals as they vary by factor group. If using by_covar, the summary of the
linear regression on the residuals by the covariate. If using both, both of
the previous tests plus a multivariate regression by the interaction of the
covariate and the factor on the residuals. See details. Summaries are named by
the dependent variable of the model and the factor and/or covariate split
for easy differentiation.
}
}
\description{
Express the residuals of an \code{lmer} model by a factor or covariate, and
test whether there may be systematic effects. One of \code{by_factor} or \code{by_covar}
(or both) must therefore be supplied.
}
\details{
If by_factor is supplied, residuals will be tested for systematic variation
using a one-way ANOVA between all factor levels. If by_covar is supplied, residuals
will be tested for systematic variation by regressing the residuals along
the covariate. If both are supplied, both tests will be done, in addition
to a multivariate regression incorporating both the factor and covariate.
}
