% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_express.R
\name{express_gaugepart.gam}
\alias{express_gaugepart.gam}
\title{Express GAM partial residuals to gauge systematic effects}
\usage{
\method{express_gaugepart}{gam}(
  model,
  comp_match = NULL,
  do_regex = TRUE,
  orig_data = NULL,
  by_factor = NULL,
  by_covar = NULL,
  covar_fit = "linear",
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  f_pal = "Dynamic",
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{mgcv::gam}.}

\item{comp_match}{Which components to match? If left NULL, all components of
the model will be selects. If supplied, works by successive forward partial
matching, splitting at commas. For instance, supplying \code{comp_match = "chla"}
will match any component with "chla" in the name. Supplying
\code{comp_match = "s,toc"} will match a component that contains "s" followed by
"toc" anywhere afterwards in the component name, e.g., "s(toc)".
Supplying \code{comp_match = "ti,chla,toc"} will then match "ti(chla, toc)".}

\item{do_regex}{Should \code{comp_match} work by selective forward partial matching?
Default TRUE.}

\item{orig_data}{Original data on which the model was built. This can either
be supplied with the original data frame directly, or, with \code{mgcv::gam}
objects, the original data can be contained within the model itself by
supplying the argument \code{control = list(keepData = TRUE)}, in which case,
orig_data can be left NULL and the data will be taken from the model object.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour. Residuals will be tested for systematic variation
using a one-way ANOVA.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size. Residuals will
be tested for systematic variation by regressing the residuals along the
covariate. See \code{covar_fit}.}

\item{covar_fit}{Specify a transform for the covariate for regression.
If there is a presumed relatonship between the covariate
and the residuals, might it be nonlinear? Default is
\code{"linear"} (i.e., no transform) but can be any of \code{c("linear", "exp", "log")}.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
regression estimate of the covariate and residuals.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_fill_discrete_qualitative} palette.}

\item{...}{Unused.}
}
\value{
A named list containing:
\itemize{
\item \code{"plots"}: At least one for each component passed to the function.
If using by_factor, a boxplot of the residuals by factor group.
If using by_covar, a scatterplot of the residuals along the covariate. If
using both, both. Plots are named by the component, the dependent variable
of the model, and either the factor or covariate split for easy differentiation.
\item \code{"summaries"} At least one for each component passed to the function.
If using by_factor, the summary of the one-way ANOVA for the
residuals as they vary by factor group. If using by_covar, the summary of the
linear regression on the residuals by the covariate. If using both, both of
the previous tests plus a linear mixed model of the covariate on the
residuals with a random effect of the factor. See details. Summaries are
named by the component, the dependent variable of the model, and the factor
and/or covariate split for easy differentiation.
}
}
\description{
Express the partial residuals of an \code{mgcv::gam} object by a factor or covariate, and
test whether there may be systematic effects. One of \code{by_factor} or \code{by_covar}
(or both) must therefore be supplied.
}
\details{
If by_factor is supplied, residuals will be tested for systematic variation
using a one-way ANOVA between all factor levels. If by_covar is supplied, residuals
will be tested for systematic variation by regressing the residuals along
the covariate. If both are supplied, both tests will be done, in addition
to a mixed effects regression incorporating both the factor and covariate.
}
