% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_gaugepart}
\alias{express_gaugepart}
\title{Express a model's partial residuals to gauge systematic effects}
\usage{
express_gaugepart(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A list of plots and model summaries, one for each model component.
For more information, see the documentation for each individual method.
}
\description{
Gauge whether there could be potential systematic variation in a model by plotting and
analyzing model partial residuals according to a factor or covariate.
Currently supported model types include:
\itemize{
\item \code{lm}: Gauge potential missing effects for a (multivariate) linear model. (\link{express_gaugepart.lm})
\item \code{glm}: Gauge potential missing effects for a (multivariate) generalized
linear model. (\link{express_gaugepart.glm})
\item \code{lmm}: Gauge potential missing effects for an \code{lme4::lmerMod} model.
(\link{express_gaugepart.lmerMod})
\item \code{glmm}: Gauge potential missing effects for an \code{lme4::glmerMod} model.
(\link{express_gaugepart.glmerMod})
\item \code{gam}: Gauge potential missing effects for an \code{mgcv::gam} model. (\link{express_gaugepart.gam})
}
}
\details{
Generic plotting and analysis of model partial residuals with
additional faceting by a factor or covariate. Uses cowplot and colorspace
along with ggplot2.
}
