% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_express.R
\name{express_hist.glm}
\alias{express_hist.glm}
\title{Express a histogram of GLM residuals}
\usage{
\method{express_hist}{glm}(
  model,
  orig_data = NULL,
  by_factor = NULL,
  res_type = "deviance",
  s_col = "black",
  r_col = "steelblue3",
  f_pal = "Dynamic",
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{glm}.}

\item{orig_data}{Original data on which the model was built. If left NULL,
it will take from the model object.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour.}

\item{res_type}{Residual type. Will work with any of the types that can be
passed to \code{residuals.glm}.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
parameter estimate.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_color_discrete_qualitative} palette.}

\item{...}{Unused.}
}
\value{
A ggplot object for a histogram of residuals, potentially faceted
by the supplied factor.
}
\description{
Express histogram of residuals for a \code{glm} object, allowing for aestheticizing by factors or
covariates potentially not included in the model definition itself.
}
\details{
Does not currently support faceting by a continuous variable. The function
will tell you if by_covar is supplied, that it cannot do this, but it will
still return a histogram (or histograms, if by_factor is also supplied).
}
