% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_hist}
\alias{express_hist}
\title{Express a model residual histogram}
\usage{
express_hist(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A ggplot object for the model residual histogram.
}
\description{
Plot a histogram of full model residuals as a ggplot object that includes
aesthetics mapped to multiple variables, whether or not they are included in
the actual definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Histogram of residuals for a (multivariate) linear model. (\link{express_hist.lm})
\item \code{glm}: Histogram of residuals for a (multivariate) generalized linear model.
(\link{express_hist.glm})
\item \code{lmm}: Histogram of residuals for an \code{lme4::lmerMod} model. (\link{express_hist.lmerMod})
\item \code{glmm}: Histogram of residuals for an \code{lme4::glmerMod} model. (\link{express_hist.glmerMod})
\item \code{gam}: Histogram of residuals for an \code{mgcv::gam} model. (\link{express_hist.gam})
}
}
\details{
Generic plotting of model residual histograms as ggplot objects with optional
additional faceting. Uses cowplot and colorspace
along with ggplot2.
}
