% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_lev}
\alias{express_lev}
\title{Express the leverage plot for a model}
\usage{
express_lev(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A ggplot object for the model residuals vs. leverage.
}
\description{
Plot the leverage vs. residuals of full model as a ggplot object that includes
aesthetics mapped to multiple variables, whether or not they are included in
the actual definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Leverage vs. residuals for a (multivariate) linear model. (\link{express_lev.lm})
\item \code{glm}: Leverage vs. residuals for a (multivariate) generalized linear model.
(\link{express_lev.glm})
}
}
\details{
Generic plotting of model residuals vs. leverage as ggplot objects with optional
additional faceting. Uses cowplot and colorspace
along with ggplot2.
}
