% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_express.R
\name{express_linpred.lmerMod}
\alias{express_linpred.lmerMod}
\title{Express a linear predictor vs. residuals plot for an LMM}
\usage{
\method{express_linpred}{lmerMod}(
  model,
  orig_data,
  by_factor = NULL,
  by_covar = NULL,
  res_type = "pearson",
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  f_pal = "Dynamic",
  cov_pal = "Sunset Dark",
  cov_pal_rev = FALSE,
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{lmerMod}.}

\item{orig_data}{Original data on which the model was built.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size.}

\item{res_type}{Residual type. Will work with any of the types that can be
passed to \code{residuals.default}.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
parameter estimate.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_color_discrete_qualitative} palette.}

\item{cov_pal}{"Covariate" palette. If mapping an aesthetic to a covariate,
which palette to use? Must be a
\code{colorspace::scale_color_continuous_sequential} palette.}

\item{cov_pal_rev}{Reverse the covariate palette? Binary TRUE/FALSE.}

\item{...}{Unused.}
}
\value{
A ggplot object of the linear predictor vs. residuals plot with
mapping to the supplied factor and/or covariate.
}
\description{
Express a linear predictor (fitted values) vs. residuals plot of an \code{lmer}
model, allowing for aestheticizing by factors or covariates potentially not
included in the model definition itself.
}
