% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_linpred}
\alias{express_linpred}
\title{Express a model linear predictor vs. residuals plot}
\usage{
express_linpred(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A ggplot object for the model linear predictor vs. residuals plot.
}
\description{
Plot the linear predictor vs. residuals for a full model as a ggplot object that includes
aesthetics mapped to multiple variables, whether or not they are included in
the actual definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Linear predictor (a.k.a. fitted values) vs. residuals for a (multivariate) linear model. (\link{express_linpred.lm})
\item \code{glm}: Linear predictor (a.k.a. fitted values) vs. residuals for a
(multivariate) generalized linear model. (\link{express_linpred.glm})
\item \code{lmm}: Linear predictor (a.k.a . fitted values) vs. residuals for
an \code{lme4::lmerMod} model. (\link{express_linpred.lmerMod})
\item \code{glmm}: Linear predictor (a.k.a . fitted values) vs. residuals for
an \code{lme4::glmerMod} model. (\link{express_linpred.glmerMod})
\item \code{gam}: Linear predictor vs. residuals plot for an \code{mgcv::gam} model. (\link{express_linpred.gam})
}
}
\details{
Generic plotting of model linear predictor vs. residual plots as ggplot objects with optional
additional faceting. Uses cowplot and colorspace
along with ggplot2.
}
