% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_express.R
\name{express_many.glm}
\alias{express_many.glm}
\title{Express many GLM components}
\usage{
\method{express_many}{glm}(
  model,
  comp_match = NULL,
  orig_data = NULL,
  by_factor = NULL,
  by_covar = NULL,
  grid = TRUE,
  b_col = "black",
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  f_pal = "Dynamic",
  cov_pal = "Sunset Dark",
  cov_pal_rev = FALSE,
  show_rug = FALSE,
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{glm}.}

\item{comp_match}{Which components to match? If left NULL, all components
will be plotted.}

\item{orig_data}{Original data on which the model was built. If left NULL,
it will take from the model object.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size.}

\item{grid}{Whether to return a combined \code{cowplot::plot_grid} object of all
the plots (default = TRUE). If FALSE, will return a list of ggplot objects.}

\item{b_col}{"Background" colour. Controls the colour of the confidence
interval around the parameter estimate.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
parameter estimate.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_color_discrete_qualitative} palette.}

\item{cov_pal}{"Covariate" palette. If mapping an aesthetic to a covariate,
which palette to use? Must be a
\code{colorspace::scale_color_continuous_sequential} palette.}

\item{cov_pal_rev}{Reverse the covariate palette? Binary TRUE/FALSE.}

\item{show_rug}{Should a "rug" be drawn on the bottom edge of the plot?}

\item{...}{Unused.}
}
\value{
A \code{cowplot::plot_grid} object of the selected components with
partial residuals mapped to the supplied factor and/or covariate, or if \code{grid = FALSE},
a list of ggplot objects.
}
\description{
Express multiple components of a \code{glm} object according to their fit-adjusted partial
residuals (component + partial residuals), allowing for aestheticizing
by factors or covariates potentially not in the model definition itself.
}
