% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_many}
\alias{express_many}
\title{Express many model components}
\usage{
express_many(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A \code{cowplot::plot_grid} object for multiple model components.
}
\description{
Plot multiple model components as a \code{cowplot::plot_grid} object that includes
aesthetics mapped to multiple variables, whether or not they are included in
the actual definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Plot selected components of a (multivariate) linear model.
Supports interaction terms. (\link{express_many.lm})
\item \code{glm}: Plot selected components of a (multivariate) generalized linear
model. Supports interaction terms. (\link{express_many.glm})
\item \code{lmm}: Plot selected components of an \code{lme4::lmerMod} model. Supports
linear and random effects. (\link{express_many.lmerMod})
\item \code{glmm}: Plot selected components of an \code{lme4::glmerMod} model. Supports
linear and random effects. (\link{express_many.glmerMod})
\item \code{gam}: Plot selected smooths of an \code{mgcv::gam} model. Supports univariate,
multivariate, and random effect smooths. (\link{express_many.gam})
}
}
\details{
Generic plotting of model components as ggplot objects with optional
additional faceting, within a \code{cowplot::plot_grid}. Uses cowplot and colorspace
along with ggplot2.
}
