% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_express.R
\name{express_one.gam}
\alias{express_one.gam}
\title{Express a single GAM smooth}
\usage{
\method{express_one}{gam}(
  model,
  comp_match,
  do_regex = TRUE,
  orig_data = NULL,
  by_factor = NULL,
  by_covar = NULL,
  b_col = "black",
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  s_pal = "Blue-Red 3",
  s_pal_rev = FALSE,
  f_pal = "Dynamic",
  cov_pal = "Sunset Dark",
  cov_pal_rev = FALSE,
  show_rug = FALSE,
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{mgcv::gam}.}

\item{comp_match}{Which component to match? Works by partial match. For
instance, supplying \code{comp_match = "chla"} will the component with "chla"
in the name. For \code{express_one}, only a single match is allowed.}

\item{do_regex}{Should \code{comp_match} work by selective forward partial matching?
Default TRUE.}

\item{orig_data}{Original data on which the model was built. This can either
be supplied with the original data frame directly, or, with \code{mgcv::gam}
objects, the original data can be contained within the model itself by
supplying the argument \code{control = list(keepData = TRUE)}, in which case,
orig_data can be left NULL and the data will be taken from the model object.}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size.}

\item{b_col}{"Background" colour. Controls the colour of the confidence
interval around the parameter estimate.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
parameter estimate.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{s_pal}{"Smooth" palette. For multidimensional smooths, the palette of
the heatmap. Must be a \code{colorspace::scale_fill_continuous_diverging} palette.}

\item{s_pal_rev}{Reverse the smooth palette? Binary TRUE/FALSE.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_color_discrete_qualitative} palette.}

\item{cov_pal}{"Covariate" palette. If mapping an aesthetic to a covariate,
which palette to use? Must be a
\code{colorspace::scale_color_continuous_sequential} palette.}

\item{cov_pal_rev}{Reverse the covariate palette? Binary TRUE/FALSE.}

\item{show_rug}{Should a "rug" be drawn on the bottom edge of the plot?
Default in \code{mgcv::plot.gam} and \code{gratia::draw.gam} is TRUE, here, it is FALSE.}

\item{...}{Unused.}
}
\value{
A ggplot object of the selected component with partial residuals
mapped to the supplied factor and/or covariate.
}
\description{
Express a single smooth for an \code{mgcv::gam} object in a similar way to
\code{mgcv::plot.gam} or \code{gratia::draw.gam}, but allow for aestheticizing by factors or
covariates potentially not included in the model definition itself.
}
