% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_express.R
\name{express_one.glmerMod}
\alias{express_one.glmerMod}
\title{Express a GLMM component}
\usage{
\method{express_one}{glmerMod}(
  model,
  orig_data,
  comp_match,
  by_factor = NULL,
  by_covar = NULL,
  b_col = "black",
  s_col = "black",
  r_col = "steelblue3",
  r_opac = 0.8,
  f_pal = "Dynamic",
  cov_pal = "Sunset Dark",
  cov_pal_rev = FALSE,
  show_rug = FALSE,
  ...
)
}
\arguments{
\item{model}{A supplied model of type \code{lme4::glmerMod}.}

\item{orig_data}{Original data on which the model was built.}

\item{comp_match}{Which component to match? Must fully match only one
component present in the model. Component names can be found in, e.g.,
\code{summary.merMod}. Note that effect terms should be referred to by their
original name, not their implementation in the model. For instance,
if there is a random effect term in a model \code{(1|Lake_ID)}, refer to this
component simply by "Lake_ID", not "(1|Lake_ID)".}

\item{by_factor}{A factor to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. by_factor
will map by colour.}

\item{by_covar}{A covariate to map aesthetics to. Must be present in the
original data. If left NULL, no aesthetic mapping is performed. On its own,
by_covar will map by a continuous colour scale. If by_factor is also present,
by_factor will map by colour and by_covar will map by size.}

\item{b_col}{"Background" colour. Controls the colour of the confidence
interval around the parameter estimate.}

\item{s_col}{"Smooth" colour. Controls the colour of the line drawn for the
parameter estimate.}

\item{r_col}{"Residual" colour. Controls the colour of the residuals drawn
for the model component.}

\item{r_opac}{"Residual" opacity. Controls the opacity of the residuals drawn
for the model component.}

\item{f_pal}{"Factor" palette. If mapping an aesthetic to a factor, which
palette to use? Must be a
\code{colorspace::scale_color_discrete_qualitative} palette.}

\item{cov_pal}{"Covariate" palette. If mapping an aesthetic to a covariate,
which palette to use? Must be a
\code{colorspace::scale_color_continuous_sequential} palette.}

\item{cov_pal_rev}{Reverse the covariate palette? Binary TRUE/FALSE.}

\item{show_rug}{Should a "rug" be drawn on the bottom edge of the plot?}

\item{...}{Unused.}
}
\value{
A ggplot object of the selected component with partial residuals
mapped to the supplied factor and/or covariate.
}
\description{
Express a linear or random component of a \code{glmer} model, allowing for aestheticizing
by factors or covariates potentially not in the model definition itself.
}
\details{
Linear terms will be plotted as their fit-adjusted partial residuals
(component + partial residuals). Random terms will be plotted as a Q-Q plot
of their effects on the model. If the random term has multiple effects, e.g.,
both an intercept and slope, the plot will be faceted for each effect.
}
