% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_one}
\alias{express_one}
\title{Express a single model component}
\usage{
express_one(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A ggplot object for a single model component.
}
\description{
Plot a singular model component as a ggplot object that includes aesthetics
mapped to multiple variables, whether or not they are included in the actual
definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Plot a selected component of a (multivariate) linear model.
Supports interaction terms. (\link{express_one.lm})
\item \code{glm}: Plot a selected component of a (multivariate) generalized linear
linear model. Supports interactions terms. (\link{express_one.glm})
\item \code{lmm}: Plot a selected component of an \code{lme4::lmerMod} model. Supports
linear and random effects. (\link{express_one.lmerMod})
\item \code{glmm}: Plot a selected component of an \code{lme4::glmerMod} model. Supports
linear and random effects. (\link{express_one.glmerMod})
\item \code{gam}: Plot a selected smooth of an \code{mgcv::gam} model. Supports univariate,
multivariate, and random effect smooths. (\link{express_one.gam})
}
}
\details{
Generic plotting of model components as ggplot objects with optional
additional faceting. Uses cowplot and colorspace along with ggplot2.
}
