% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_qqresid}
\alias{express_qqresid}
\title{Express a quantile-quantile plot of model residuals}
\usage{
express_qqresid(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A ggplot object for the model residual q-q plot.
}
\description{
Plot the quantile-quantile plot of full model residuals as a ggplot object that includes
aesthetics mapped to multiple variables, whether or not they are included in
the actual definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Residual Q-Q plot for a (multivariate) linear model. (\link{express_qqresid.lm})
\item \code{glm}: Residual Q-Q plot for a (multivariate) generalized linear model. (\link{express_qqresid.glm})
\item \code{lmm}: Residual Q-Q plot for an \code{lme4::lmerMod} model. (\link{express_qqresid.lmerMod})
\item \code{glmm}: Residual Q-Q plot for an \code{lme4::glmerMod} model. (\link{express_qqresid.glmerMod})
\item \code{gam}: Residual Q-Q plot for an \code{mgcv::gam} model. (\link{express_qqresid.gam})
}
}
\details{
Generic plotting of model residual Q-Q plots as ggplot objects with optional
additional faceting. Uses cowplot and colorspace
along with ggplot2.
}
