% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_express.R
\name{express_scloc}
\alias{express_scloc}
\title{Express the scale-location plot for a model}
\usage{
express_scloc(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to specific methods based on model type.}
}
\value{
A ggplot object for the model scale-locations.
}
\description{
Plot the scale-location of full model (fitted values vs. transformed residuals) as a ggplot object that includes
aesthetics mapped to multiple variables, whether or not they are included in
the actual definition of the model.
Currently supported model types include:
\itemize{
\item \code{lm}: Scale-location plot for a (multivariate) linear model. (\link{express_scloc.lm})
\item \code{glm}: Scale-location plot for a (multivariate) generalized linear model.
(\link{express_scloc.lm})
}
}
\details{
Generic plotting of model scale-locations as ggplot objects with optional
additional faceting. Uses cowplot and colorspace
along with ggplot2.
}
